Revert "Move `Attr` From Cython to Pure Python (#1326)"
* This reverts commit bdd4bdbf0c5720c31415f2bcb8badfb5232794d2.
* There are issues with the lifetime of `Context` in multiprocess
  settings as observed in cloud Jupyter notebooks and dask.
