+++ tiledb/cc/enum.cc
@@ -20,21 +20,21 @@ void init_enums(py::module &m) {
   m.def("TILEDB_OFFSET_SIZE", []() { return TILEDB_OFFSET_SIZE; });
   m.def("TILEDB_TIMESTAMP_NOW_MS", []() { return TILEDB_TIMESTAMP_NOW_MS; });
 
-  py::enum_<tiledb_datatype_t>(m, "DataType", py::module_local()) DENUM(
-      INT32) DENUM(INT64) DENUM(FLOAT32) DENUM(FLOAT64) DENUM(CHAR) DENUM(INT8)
-      DENUM(UINT8) DENUM(INT16) DENUM(UINT16) DENUM(UINT32) DENUM(UINT64) DENUM(
-          BOOL) DENUM(STRING_ASCII) DENUM(STRING_UTF8) DENUM(STRING_UTF16)
-          DENUM(STRING_UTF32) DENUM(STRING_UCS2) DENUM(STRING_UCS4) DENUM(ANY)
-              DENUM(DATETIME_YEAR) DENUM(DATETIME_MONTH) DENUM(DATETIME_WEEK)
-                  DENUM(DATETIME_DAY) DENUM(DATETIME_HR) DENUM(DATETIME_MIN)
-                      DENUM(DATETIME_SEC) DENUM(DATETIME_MS) DENUM(DATETIME_US)
-                          DENUM(DATETIME_NS) DENUM(DATETIME_PS)
-                              DENUM(DATETIME_FS) DENUM(DATETIME_AS)
-                                  DENUM(TIME_HR) DENUM(TIME_MIN) DENUM(TIME_SEC)
-                                      DENUM(TIME_MS) DENUM(TIME_US)
-                                          DENUM(TIME_NS) DENUM(TIME_PS)
-                                              DENUM(TIME_FS) DENUM(TIME_AS)
-                                                  DENUM(BLOB);
+  py::enum_<tiledb_datatype_t>(m, "DataType", py::module_local()) DENUM(INT32)
+      DENUM(INT64) DENUM(FLOAT32) DENUM(FLOAT64) DENUM(CHAR) DENUM(INT8)
+          DENUM(UINT8) DENUM(INT16) DENUM(UINT16) DENUM(UINT32) DENUM(UINT64)
+              DENUM(STRING_ASCII) DENUM(STRING_UTF8) DENUM(STRING_UTF16) DENUM(
+                  STRING_UTF32) DENUM(STRING_UCS2) DENUM(STRING_UCS4) DENUM(ANY)
+                  DENUM(DATETIME_YEAR) DENUM(DATETIME_WEEK) DENUM(DATETIME_DAY)
+                      DENUM(DATETIME_HR) DENUM(DATETIME_MIN) DENUM(DATETIME_SEC)
+                          DENUM(DATETIME_MS) DENUM(DATETIME_US)
+                              DENUM(DATETIME_NS) DENUM(DATETIME_PS)
+                                  DENUM(DATETIME_FS) DENUM(DATETIME_AS)
+                                      DENUM(TIME_HR) DENUM(TIME_MIN)
+                                          DENUM(TIME_SEC) DENUM(TIME_MS)
+                                              DENUM(TIME_US) DENUM(TIME_NS)
+                                                  DENUM(TIME_PS) DENUM(TIME_FS)
+                                                      DENUM(TIME_AS);
 
   py::enum_<tiledb_array_type_t>(m, "ArrayType") DENUM(DENSE) DENUM(SPARSE);
