--- tiledb/tests/test_attribute.py
@@ -1,179 +1,0 @@
-import xml.etree.ElementTree
-
-import numpy as np
-import pytest
-from numpy.testing import assert_array_equal
-
-import tiledb
-from tiledb.tests.common import assert_captured, DiskTestCase, has_pandas
-
-
-class AttributeTest(DiskTestCase):
-    def test_minimal_attribute(self):
-        attr = tiledb.Attr()
-        self.assertTrue(attr.isanon)
-        self.assertEqual(attr.name, "")
-        self.assertEqual(attr.dtype, np.float_)
-        # self.assertEqual(attr.compressor, (None, -1))
-        self.assertFalse(attr.isvar)
-        self.assertFalse(attr.isnullable)
-
-        try:
-            assert xml.etree.ElementTree.fromstring(attr._repr_html_()) is not None
-        except:
-            pytest.fail(f"Could not parse attr._repr_html_(). Saw {attr._repr_html_()}")
-
-    def test_attribute(self, capfd):
-        attr = tiledb.Attr("foo")
-
-        attr.dump()
-        assert_captured(capfd, "Name: foo")
-
-        assert attr.name == "foo"
-        assert attr.dtype == np.float64, "default attribute type is float64"
-        # compressor, level = attr.compressor
-        # self.assertEqual(compressor, None, "default to no compression")
-        # self.assertEqual(level, -1, "default compression level when none is specified")
-
-    @pytest.mark.parametrize(
-        "dtype, fill",
-        [
-            (np.dtype(bytes), b"abc"),
-            # (str, "defg"),
-            (np.float32, np.float32(0.4023573667780681)),
-            (np.float64, np.float64(0.0560602549760851)),
-            (np.dtype("M8[ns]"), np.timedelta64(11, "ns")),
-            (np.dtype([("f0", "<i4"), ("f1", "<i4"), ("f2", "<i4")]), (1, 2, 3)),
-        ],
-    )
-    def test_attribute_fill(self, dtype, fill):
-        attr = tiledb.Attr("", dtype=dtype, fill=fill)
-        assert np.array(attr.fill, dtype=dtype) == np.array(fill, dtype=dtype)
-
-        path = self.path()
-        dom = tiledb.Domain(tiledb.Dim(domain=(0, 0), tile=1, dtype=np.int64))
-        schema = tiledb.ArraySchema(domain=dom, attrs=(attr,))
-        tiledb.DenseArray.create(path, schema)
-
-        with tiledb.open(path) as R:
-            assert R.multi_index[0][""] == np.array(fill, dtype=dtype)
-            assert R[0] == np.array(fill, dtype=dtype)
-            if has_pandas() and not hasattr(dtype, "fields"):
-                # record type unsupported for .df
-                assert R.df[0][""].values == np.array(fill, dtype=dtype)
-
-    def test_full_attribute(self, capfd):
-        filter_list = tiledb.FilterList([tiledb.ZstdFilter(10)])
-        filter_list = tiledb.FilterList([tiledb.ZstdFilter(10)])
-        attr = tiledb.Attr("foo", dtype=np.int64, filters=filter_list)
-
-        attr.dump()
-        assert_captured(capfd, "Name: foo")
-
-        self.assertEqual(attr.name, "foo")
-        self.assertEqual(attr.dtype, np.int64)
-
-        # <todo>
-        # compressor, level = attr.compressor
-        # self.assertEqual(compressor, "zstd")
-        # self.assertEqual(level, 10)
-
-    def test_ncell_attribute(self):
-        dtype = np.dtype([("", np.int32), ("", np.int32), ("", np.int32)])
-        attr = tiledb.Attr("foo", dtype=dtype)
-
-        self.assertEqual(attr.dtype, dtype)
-        self.assertEqual(attr.ncells, 3)
-
-        # dtype subarrays not supported
-        with self.assertRaises(TypeError):
-            tiledb.Attr("foo", dtype=np.dtype((np.int32, 2)))
-
-        # mixed type record arrays not supported
-        with self.assertRaises(TypeError):
-            tiledb.Attr("foo", dtype=np.dtype([("", np.float32), ("", np.int32)]))
-
-    def test_ncell_bytes_attribute(self):
-        dtype = np.dtype((np.bytes_, 10))
-        attr = tiledb.Attr("foo", dtype=dtype)
-
-        self.assertEqual(attr.dtype, dtype)
-        self.assertEqual(attr.ncells, 10)
-
-    def test_bytes_var_attribute(self):
-        with pytest.warns(DeprecationWarning, match="Attr given `var=True` but"):
-            attr = tiledb.Attr("foo", var=True, dtype="S1")
-            self.assertEqual(attr.dtype, np.dtype("S"))
-            self.assertTrue(attr.isvar)
-
-        with pytest.warns(DeprecationWarning, match="Attr given `var=False` but"):
-            attr = tiledb.Attr("foo", var=False, dtype="S")
-            self.assertEqual(attr.dtype, np.dtype("S"))
-            self.assertTrue(attr.isvar)
-
-        attr = tiledb.Attr("foo", var=True, dtype="S")
-        self.assertEqual(attr.dtype, np.dtype("S"))
-        self.assertTrue(attr.isvar)
-
-        attr = tiledb.Attr("foo", var=False, dtype="S1")
-        self.assertEqual(attr.dtype, np.dtype("S1"))
-        self.assertFalse(attr.isvar)
-
-        attr = tiledb.Attr("foo", dtype="S1")
-        self.assertEqual(attr.dtype, np.dtype("S1"))
-        self.assertFalse(attr.isvar)
-
-        attr = tiledb.Attr("foo", dtype="S")
-        self.assertEqual(attr.dtype, np.dtype("S"))
-        self.assertTrue(attr.isvar)
-
-    def test_nullable_attribute(self):
-        attr = tiledb.Attr("nullable", nullable=True, dtype=np.int32)
-        self.assertEqual(attr.dtype, np.dtype(np.int32))
-        self.assertTrue(attr.isnullable)
-
-    def test_datetime_attribute(self):
-        attr = tiledb.Attr("foo", dtype=np.datetime64("", "D"))
-        assert attr.dtype == np.dtype(np.datetime64("", "D"))
-        assert attr.dtype != np.dtype(np.datetime64("", "Y"))
-        assert attr.dtype != np.dtype(np.datetime64)
-
-    @pytest.mark.parametrize("sparse", [True, False])
-    def test_ascii_attribute(self, sparse, capfd):
-        path = self.path("test_ascii")
-        dom = tiledb.Domain(
-            tiledb.Dim(name="d", domain=(1, 4), tile=1, dtype=np.uint32)
-        )
-
-        with pytest.raises(TypeError) as exc_info:
-            tiledb.Attr(name="A", dtype="ascii", var=False)
-        assert (
-            str(exc_info.value) == "dtype is not compatible with var-length attribute"
-        )
-
-        attrs = [tiledb.Attr(name="A", dtype="ascii")]
-
-        schema = tiledb.ArraySchema(domain=dom, attrs=attrs, sparse=sparse)
-        tiledb.Array.create(path, schema)
-
-        ascii_data = ["a", "b", "c", "ABC"]
-        unicode_data = ["±", "×", "÷", "√"]
-
-        with tiledb.open(path, "w") as A:
-            if sparse:
-                with self.assertRaises(tiledb.TileDBError):
-                    A[np.arange(1, 5)] = unicode_data
-                A[np.arange(1, 5)] = ascii_data
-            else:
-                with self.assertRaises(tiledb.TileDBError):
-                    A[:] = unicode_data
-                A[:] = ascii_data
-
-        with tiledb.open(path, "r") as A:
-            assert A.schema.nattr == 1
-            A.schema.dump()
-            assert_captured(capfd, "Type: STRING_ASCII")
-            assert A.schema.attr("A").isvar
-            assert A.schema.attr("A").dtype == np.bytes_
-            assert A.schema.attr("A").isascii
-            assert_array_equal(A[:]["A"], np.asarray(ascii_data, dtype=np.bytes_))
